// models/project/src/mc-project/applied-data-fence/types.ts
import type { TBuilder } from '@commercetools-test-data/core';

export type TAppliedDataFence = {
  type: string;
  group: string;
  name: string;
  value: boolean;
};

export type TAppliedDataFenceGraphql = TAppliedDataFence & {
  __typename: 'AppliedDataFence';
};

export type TAppliedDataFenceBuilder = TBuilder<TAppliedDataFence>;
export type TCreateAppliedDataFenceBuilder = () => TAppliedDataFenceBuilder;